- Packet logList working; - Message logList Working; - Themed; - All buttons working; - Dump to log file.
