--- src/main/kotlin/components/DisplayComponents.kt
@@ -1,67 +1,0 @@
-package components
-
-import androidx.compose.foundation.background
-import androidx.compose.foundation.border
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.Surface
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.MutableState
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.input.TextFieldValue
-import androidx.compose.ui.text.style.TextOverflow
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-
-@Composable
-fun DisplayPanel(
-    modifier: Modifier,
-    mainOutput: MutableState<TextFieldValue>
-) {
-    Column(
-        modifier = modifier
-            .fillMaxWidth()
-            .background(MaterialTheme.colors.surface)
-            .background(Color.White)
-            .border(color = Color.Gray, width = 1.dp)
-            .padding(start = 16.dp, end = 16.dp),
-        verticalArrangement = Arrangement.Center,
-        horizontalAlignment = Alignment.End
-    ) {
-        Text(
-            text = mainOutput.value.text,
-            style = TextStyle(
-                fontSize = 48.sp
-            ),
-            overflow = TextOverflow.Ellipsis,
-            softWrap = false,
-            maxLines = 1,
-        )
-    }
-}
-
-@Composable
-fun Keyboard(
-    modifier: Modifier,
-    mainOutput: MutableState<TextFieldValue>
-) {
-    Surface(modifier) {
-        KeyboardKeys(mainOutput)
-    }
-}
-
-@Composable
-fun KeyboardKeys(mainOutput: MutableState<TextFieldValue>) {
-    Row(modifier = Modifier.fillMaxSize()) {
-        Text("New Text HERE !")
-    }
-}
