+++ src/main/kotlin/main.kt
@@ -1,71 +1,49 @@
 import androidx.compose.desktop.Window
-import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.padding
 import androidx.compose.material.MaterialTheme
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
+import androidx.compose.material.darkColors
+import androidx.compose.material.lightColors
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.unit.IntSize
-import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import components.center
 import components.leftSide
-import components.rightSide
 import components.top
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.channels.Channel
-import kotlinx.coroutines.channels.SendChannel
-import kotlinx.coroutines.launch
-import kotlinx.coroutines.runBlocking
-import model.Log
-import services.LogWriter
+import services.LogWriterService
 import services.ProxyService
-import kotlin.concurrent.thread
 
-const val DEFAULT_WIDTH = 1250
+const val DEFAULT_WIDTH = 1300
 const val DEFAULT_HEIGHT = 930
 val OVER_LINE_STYLE = TextStyle(
     fontWeight = FontWeight.Normal,
     fontSize = 13.sp,
     letterSpacing = 1.5.sp
 )
-val logWriter = LogWriter()
+val logWriterService = LogWriterService()
 
 fun main() = Window(
     title = "DProxyN - Sniffer Tunneling",
     size = IntSize(DEFAULT_WIDTH, DEFAULT_HEIGHT),
     resizable = false
 ) {
-    val proxyService = ProxyService(logWriter)
+    val proxyService = ProxyService(logWriterService)
 
     MaterialTheme {
 
-        Column {
-            top()
-        }
+
         Row {
             leftSide(
                 modifier = Modifier,
                 proxyService = proxyService,
-                logWriter = logWriter
+                logWriterService = logWriterService
             )
-            center(
-                modifier = Modifier,
-                logWriter = logWriter
-            )
-            rightSide(
-                modifier = Modifier,
-                proxyService = proxyService,
-                logWriter = logWriter
-            )
+        }
+        Column {
+            top()
         }
     }
 }
