+++ src/main/kotlin/services/ProxyService.kt
@@ -5,8 +5,8 @@ import java.io.IOException
 import java.net.Socket
 
 class ProxyService(
-    private val logWriter: LogWriter,
-    private val sendReceive: SendReceiveService = SendReceiveService(logWriter)
+    private val logWriterService: LogWriterService,
+    private val communicationService: CommunicationService = CommunicationService(logWriterService)
 ) {
     var serverLog = true
     var clientLog = true
@@ -17,10 +17,11 @@ class ProxyService(
     fun setConnections(localPort: Int, remotePort: Int, remoteAddress: String) {
 
         if (localPort == 0 || remotePort == 0) {
+            logWriterService.systemLog("Fail to start: Ports must not be null")
             throw IOException("Ports must not be null")
         }
-        localConnection = SocketConfig(logWriter).openLocalConnection(localPort)
-        remoteConnection = SocketConfig(logWriter).openRemoteConnection(remoteAddress, remotePort)
+        localConnection = SessionSetupService(logWriterService).openLocalConnection(localPort)
+        remoteConnection = SessionSetupService(logWriterService).openRemoteConnection(remoteAddress, remotePort)
 
         runProxy(localConnection = localConnection, remoteConnection = remoteConnection)
 
@@ -32,7 +33,7 @@ class ProxyService(
         do {
             // --- Server Server ( receive packets from server )
             val serverPackets =
-                sendReceive.receive(
+                communicationService.receive(
                     connectServer = remoteConnection,
                     indicator = Indicator.Server,
                     log = serverLog
@@ -41,7 +42,7 @@ class ProxyService(
             //serverPackets = filter(serverPackets)
 
             // --- Server Client
-            sendReceive.send(
+            communicationService.send(
                 connectServer = localConnection,
                 packetToSend = serverPackets,
                 indicator = Indicator.Client,
@@ -49,14 +50,14 @@ class ProxyService(
             )
 
             // --- Client Server
-            val clientPackets = sendReceive.receive(
+            val clientPackets = communicationService.receive(
                 connectServer = localConnection,
                 indicator = Indicator.Client,
                 log = clientLog
             )
 
             // --- Server Server ( send packets to server )
-            sendReceive.send(
+            communicationService.send(
                 connectServer = remoteConnection,
                 packetToSend = clientPackets,
                 indicator = Indicator.Server,
@@ -69,12 +70,12 @@ class ProxyService(
             localConnection.close()
             remoteConnection.close()
         }
-        logWriter.systemLog("Disconnected from the server")
+        logWriterService.systemLog("Disconnected from the server")
     }
 
     fun sendPacket2Server(packet: String, packetNumber: Long) {
         try {
-            sendReceive.send(
+            communicationService.send(
                 connectServer = remoteConnection,
                 packetToSend = packet,
                 indicator = Indicator.iServer,
@@ -82,13 +83,13 @@ class ProxyService(
                 iPacketNumber = packetNumber
             )
         }catch (ex: Exception) {
-            logWriter.systemLog("Failed to inject: ${ex.message}")
+            logWriterService.systemLog("Failed to inject: ${ex.message}")
         }
     }
 
     fun sendPacket2Client(packet: String, packetNumber: Long) {
         try {
-            sendReceive.send(
+            communicationService.send(
                 connectServer = localConnection,
                 packetToSend = packet,
                 indicator = Indicator.iClient,
@@ -96,7 +97,7 @@ class ProxyService(
                 iPacketNumber = packetNumber
             )
         }catch (ex: Exception) {
-            logWriter.systemLog("Failed to inject: ${ex.message}")
+            logWriterService.systemLog("Failed to inject: ${ex.message}")
         }
     }
 }
